requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu
pymupdf
langgraph